+++ bolero/optimizer/__init__.py
@@ -6,6 +6,7 @@ from .reps import REPSOptimizer
 from .creps import CREPSOptimizer
 from .acmes import ACMESOptimizer
 from .ccmaes import CCMAESOptimizer
+from .nes import XNESOptimizer
 from .cem import CEMOptimizer
 
 
@@ -23,6 +24,7 @@ __all__ = [
     "CCMAESOptimizer",
     "REPSOptimizer",
     "CREPSOptimizer",
+    "XNESOptimizer",
     "CEMOptimizer"]
 
 from .skoptimize import SkOptOptimizer, skopt_available